dz 04  done + refactor
