--- src/main/java/lesson02/Animal.java
@@ -1,24 +1,0 @@
-package lesson02;
-
-public class Animal {
-   protected String name;
-   protected int runValue;
-    protected int swimValue;
-    protected int jumpValue;
-    public Animal(String name, int runValue, int swimValue, int jumpValue){
-        this.name = name;
-        this.runValue = runValue;
-        this.swimValue = swimValue;
-        this.jumpValue = jumpValue;
-
-    }
-    public void run (int distance){
-
-    }
-    public void swim (int distance){
-
-    }
-    public void jump (int height){
-
-    }
-}
