--- src/main/java/lesson02/Cat.java
@@ -1,7 +1,0 @@
-package lesson02;
-
-public class Cat extends Animal{
-    public Cat(String name, int runValue, int swimValue, int jumpValue) {
-        super(name, runValue, swimValue, jumpValue);
-    }
-}
