--- src/main/java/lesson02/Dog.java
@@ -1,7 +1,0 @@
-package lesson02;
-
-public class Dog extends Animal{
-    public Dog(String name, int runValue, int swimValue, int jumpValue) {
-        super(name, runValue, swimValue, jumpValue);
-    }
-}
